refactor(FakerError): rename file name
