+++ src/errors/faker-error.ts
